fix: Replace deprecated boxed-primitive constructors, prefer primitives over valueOf()
- Replace `new Integer/Long/Float/Double/Boolean/Short/Byte/Character(...)`
  (deprecated for removal in modern JDKs) with primitives or `.valueOf()`
  where boxing is actually needed.
- Where the target already accepted a primitive or the value only needed
  autoboxing, drop the explicit `.valueOf()`/boxing entirely (e.g.
  `Boolean.valueOf(font.isBold())` -> `font.isBold()`,
  `Integer.valueOf(getRadius().intValue())` -> `getRadius()`).
- Fix resulting `assertEquals` overload ambiguity (int vs Object) by
  unboxing the actual value with `.intValue()` where both sides needed
  to be primitive.

+++ src/test/java/ar/com/fdvs/dj/test/crosstab/CrosstabCustomIncrementerTest.java
@@ -102,8 +102,7 @@ public class CrosstabCustomIncrementerTest extends BaseDjReportTest {
 						return null;
 					}
 
-					return Float.valueOf(
-							(initial == null ? 0 : initial.floatValue()) + (value == null ? 0 : value.floatValue()));
+					return (initial == null ? 0 : initial.floatValue()) + (value == null ? 0 : value.floatValue());
 				}
 			};
 		}
@@ -210,10 +209,10 @@ public class CrosstabCustomIncrementerTest extends BaseDjReportTest {
 		// 1 entry
 		List<Product> col = new ArrayList<Product>(2);
 
-		col.add(new Product(Long.valueOf(1), "book", "Harry Potter 7", "Florida", "Main Street", Long.valueOf(2500),
+		col.add(new Product(1L, "book", "Harry Potter 7", "Florida", "Main Street", 2500L,
 				null));
-		col.add(new Product(Long.valueOf(1), "book", "Harry Potter 7", "Florida", "Railway Station", Long.valueOf(1400),
-				Float.valueOf(2831.32f)));
+		col.add(new Product(1L, "book", "Harry Potter 7", "Florida", "Railway Station", 1400L,
+				2831.32f));
 
 		return col;
 	}
